Moved theme:android-util into :core domain
