--- application/common/theme/android-util/src/main/AndroidManifest.xml
@@ -1,4 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<manifest package="com.stepango.blockme.common.theme.android.util">
-
-</manifest>
