+++ application/core/theme/android-util/.gitignore
