+++ application/core/theme/android-util/build.gradle.kts
@@ -1,5 +1,5 @@
 androidUtil(
-    packageName = "com.stepango.blockme.common.theme.android.util",
+    packageName = "com.stepango.blockme.core.theme.android.util",
 
     dependencies = deps(
         androidx.appcompat,
