+++ application/core/theme/android-util/proguard-rules.pro
