+++ application/core/theme/android-util/src/main/java/com/stepango/blockme/core/theme/android/util/ThemeUtilsImpl.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.stepango.blockme.common.theme.android.util
+package com.stepango.blockme.core.theme.android.util
 
 import android.content.Context
 import android.content.res.Configuration
