+++ application/core/theme/android-util/src/main/java/com/stepango/blockme/core/theme/android/util/di/ThemeModule.kt
@@ -14,10 +14,10 @@
  * limitations under the License.
  */
 
-package com.stepango.blockme.common.theme.android.util.di
+package com.stepango.blockme.core.theme.android.util.di
 
-import com.stepango.blockme.common.theme.android.util.ThemeUtils
-import com.stepango.blockme.common.theme.android.util.ThemeUtilsImpl
+import com.stepango.blockme.core.theme.android.util.ThemeUtils
+import com.stepango.blockme.core.theme.android.util.ThemeUtilsImpl
 import dagger.Binds
 import dagger.Module
 import javax.inject.Singleton
