+++ application/root-library/src/main/java/com/stepango/blockme/root/library/di/RootComponent.kt
@@ -16,7 +16,7 @@
 
 package com.stepango.blockme.root.library.di
 
-import com.stepango.blockme.common.theme.android.util.di.ThemeComponent
+import com.stepango.blockme.core.theme.android.util.di.ThemeComponent
 import com.stepango.blockme.core.di.library.scopes.AppScope
 import com.stepango.blockme.root.library.SampleApp
 import dagger.Component
