+++ application/settings.gradle.kts
@@ -5,13 +5,13 @@ include(
     ":core:di:library",
     ":core:network:library",
     ":core:mvvm:library",
+    ":core:theme:android-util",
 
     ":common:extensions:android-util",
     ":common:extensions:util",
     ":common:extensions:databinding-adapter",
     ":common:progressbar:databinding",
     ":common:recyclerview:widget",
-    ":common:theme:android-util",
     ":common:util",
 
     ":root-library",
